bookkeeping/requirement.txt
django==4.1.7
djangorestframework==3.14.0
django-debug-toolbar==4.0.0
drf-yasg==1.21.5